Added update and delete methods.

+++ project/app/views.py
@@ -1,6 +1,7 @@
 from project.app import app, db
 from flask import jsonify, abort, make_response, request, url_for, render_template
 from project.app.models import Tasks
+from uuid import uuid4
 
 API_URI = '/todo/api/v1.0/tasks'
 
@@ -22,13 +23,15 @@ def get_tasks():
     if request.method == 'POST':
         post_data = request.get_json()
         new_task = Tasks(
-            title=post_data['title'],
-            description=post_data['description'],
-            done=False
+            id=uuid4().hex,
+            title=post_data.get('title'),
+            description=post_data.get('description'),
+            done=post_data.get('done')
         )
         db.session.add(new_task)
         db.session.commit()
         response_object['message'] = 'Task added!'
+        print(post_data)
     else:
         tasks = Tasks.query.all()
         output = []
@@ -41,75 +44,35 @@ def get_tasks():
             }
             output.append(task_data)
         response_object['tasks'] = output
-    # return jsonify({'tasks': list(map(make_public_task, output))})
     return jsonify(response_object)
 
 
-@app.route(API_URI + '/<int:task_id>', methods=['GET'])
-def get_task(task_id):
+@app.route('/tasks/<task_id>', methods=['PUT', 'DELETE'])
+def single_task(task_id):
+    response_object = {'status': 'success'}
     task = Tasks.query.filter_by(id=task_id).first()
-    # task = list(filter(lambda t: t['id'] == task_id, tasks))
-    if not task:
-        abort(404)
-    return jsonify({'task': task.as_dict()})
-
-
-@app.route(API_URI, methods=['POST'])
-def create_task():
-    if not request.json or 'title' not in request.json:
-        abort(400)
-    content = request.json
-    new_task = Tasks(
-        title=content['title'],
-        description=content['description'],
-        done=False
-    )
-    db.session.add(new_task)
-    db.session.commit()
-    return jsonify({'task': new_task.as_dict()}), 201
-
-
-@app.route(API_URI + '/<int:task_id>', methods=['PUT'])
-def update_task(task_id):
-    task = Tasks.query.filter_by(id=task_id).first()
-    content = request.json
-    if not task:
-        abort(404)
-    if not content:
-        abort(400)
-    if 'done' in request.json and type(request.json['done']) is not bool:
-        abort(400)
-    if 'title' in content:
-        task.title = content['title']
-    if 'description' in content:
-        task.description = content['description']
-    if 'done' in content:
-        task.done = content['done']
-    db.session.commit()
-    return jsonify({'task': task.as_dict()})
-
-
-@app.route(API_URI + '/<int:task_id>', methods=['DELETE'])
-def delete_task(task_id):
-    task = Tasks.query.filter_by(id=task_id).first()
-    if not task:
-        abort(404)
-    del task
-    Tasks.query.filter_by(id=task_id).delete()
-    db.session.commit()
-    return jsonify({'result': True})
+    if request.method == 'PUT':
+        post_data = request.get_json()
+        title = post_data.get('title')
+        description = post_data.get('description')
+        done = post_data.get('done')
+        if title:
+            task.title = title
+        if description:
+            task.description = description
+        if done is not None:
+            task.done = done
+        db.session.commit()
+        response_object['message'] = 'Task updated!'
+    if request.method == 'DELETE':
+        if not task:
+            abort(404)
+        Tasks.query.filter_by(id=task_id).delete()
+        db.session.commit()
+        response_object['message'] = 'Task removed!'
+    return jsonify(response_object)
 
 
 @app.errorhandler(404)
 def not_found(error):
     return make_response(jsonify({'error': 'Not found'}), 404)
-
-
-def make_public_task(task):
-    new_task = {}
-    for field in task:
-        if field == 'id':
-            new_task['uri'] = url_for('get_task', task_id=task['id'], _external=True)
-        else:
-            new_task[field] = task[field]
-    return new_task
